fix: keep React contexts Fast Refresh-safe to stop stale-context errors
BottomSheetCTA and TodayMenuId each colocated their `createContext`,
provider component, and consumer hook in one file. A file that exports
both a component and non-components can't use Fast Refresh, so an HMR
update re-executes the module and creates a *new* context object while
already-mounted providers/consumers still hold the old one — the
consumer's `useContext` then returns null and throws "must be used
within a provider" even though it is correctly wrapped.

Split each into a component-free module holding the context + hook (its
identity now survives HMR) and a provider-only file (which Fast Refresh
can handle). Barrels re-export both, so consumers are unchanged. This
also removes the `react-refresh/only-export-components` disables — the
warning is now resolved structurally rather than suppressed.
